WIP: Removed Newlib stubs (NOTE: This version contains no usable user system calls. Delays are made due to development roadblocks.)
- Fixed misaligned registers in `syscalls_registers_x86_32_t`
- Added `arg4` and `arg5` to `farix_syscall`
- Deleted newlib stubs

+++ arch/x86_32/bus/pci.c
@@ -18,6 +18,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 -----------------------------------------------------------------------
 */
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "hal.h"
@@ -27,7 +28,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #include "cpu/pci.h"
 
 pci_device_t pci_devices[32];
-int pci_device_count;
+size_t pci_device_count;
 
 /* Initialise the PCI by iterating through and storing all found devices into `pci_devices`. */
 void init_pci() {
